implement gate checks on the backend

+++ app/Http/Controllers/StudentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Classes;
 use App\Models\Student;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Resources\ClassesResource;
 use App\Http\Resources\StudentResource;
 use App\Http\Requests\StoreStudentRequest;
@@ -14,6 +15,8 @@ class StudentController extends Controller
 {
     public function index(Request $request)
     {
+        Gate::authorize('student_access');
+
         $studentsQuery = Student::query();
 
         $this->applySearch($studentsQuery, $request->search);
@@ -38,6 +41,8 @@ class StudentController extends Controller
 
     public function create()
     {
+        Gate::authorize('student_create');
+
         $classes = ClassesResource::collection(Classes::all());
 
         return inertia('Students/Create', [
@@ -47,6 +52,8 @@ class StudentController extends Controller
 
     public function store(StoreStudentRequest $request)
     {
+        Gate::authorize('student_create');
+
         Student::create($request->validated());
 
         return redirect()->route('students.index');
@@ -54,6 +61,8 @@ class StudentController extends Controller
 
     public function edit(Student $student)
     {
+        Gate::authorize('student_edit');
+
         $classes = ClassesResource::collection(Classes::all());
 
         return inertia('Students/Edit', [
@@ -64,6 +73,8 @@ class StudentController extends Controller
 
     public function update(UpdateStudentRequest $request, Student $student)
     {
+        Gate::authorize('student_edit');
+
         $student->update($request->validated());
 
         return redirect()->route('students.index');
@@ -71,6 +82,8 @@ class StudentController extends Controller
 
     public function destroy(Student $student)
     {
+        Gate::authorize('student_delete');
+
         $student->delete();
 
         return redirect()->route('students.index');
